# openapi.yaml
openapi: 3.0.3
info:
  title: GUARDIAN API
  description: |
    GUARDIAN Pharmaceutical Compliance Analysis API
    
    A comprehensive REST API for pharmaceutical protocol compliance analysis against European Pharmacopoeia standards.
    The system provides document processing, vector similarity search, LLM-powered compliance analysis, and professional report generation.
    
    ## Features
    
    - **Protocol Analysis**: AI-powered compliance analysis against European Pharmacopoeia standards
    - **Document Management**: Upload, process, and manage reference documents
    - **Vector Search**: Similarity search across pharmacopoeia sections
    - **Report Generation**: Professional PDF and HTML reports with visualizations
    - **Batch Processing**: Handle multiple protocols and generate bulk reports
    
    ## Authentication
    
    Currently, the API operates without authentication in development mode. 
    Production deployments should implement appropriate authentication mechanisms.
    
    ## Rate Limiting
    
    API endpoints are rate-limited to ensure system stability:
    - Standard endpoints: 100 requests/minute
    - Analysis endpoints: 10 requests/minute
    - Upload endpoints: 20 requests/minute
    
    ## Error Handling
    
    The API uses standard HTTP status codes and returns detailed error information in JSON format.
    
  version: 1.0.0
  contact:
    name: GUARDIAN Support
    email: guardian-support@qmul.ac.uk
...
  license:
    name: MIT
...
servers:
...
paths:
...
  /api/documents/upload:
    post:
      tags: [Documents]
      summary: Upload reference document
      description: |
        Uploads a reference document (PDF, DOCX, or TXT) for processing and indexing.
        Supported formats: PDF, DOCX, TXT. Maximum file size: 100MB.
      operationId: uploadDocument
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/DocumentUploadRequest'
      responses:
        '201':
          description: Document uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentUploadResponse'
        '400':
          description: Invalid file or request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/documents:
    get:
      tags: [Documents]
      summary: List documents
      description: Retrieves a paginated list of uploaded documents
      operationId: listDocuments
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 20
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, processing, completed, failed]
      responses:
        '200':
          description: Documents retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentListResponse'
  /api/documents/{doc_id}:
    get:
      tags: [Documents]
      summary: Get document information
      description: Retrieves detailed information about a specific document
      operationId: getDocument
      parameters:
        - name: doc_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Document information retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentResponse'
        '404':
          description: Document not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags: [Documents]
      summary: Delete document
      description: Deletes a document and removes it from the vector index
      operationId: deleteDocument
      parameters:
        - name: doc_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Document deleted successfully
        '404':
          description: Document not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/documents/{doc_id}/process:
    post:
      tags: [Documents]
      summary: Process document
      description: Initiates processing of an uploaded document for indexing
      operationId: processDocument
      parameters:
        - name: doc_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '202':
          description: Document processing initiated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProcessingResponse'

  # Search Endpoints
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error type identifier
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          description: Additional error details
        timestamp:
          type: string
          format: date-time
          description: Error timestamp

    # Health Schemas
...
    DocumentUploadRequest:
      type: object
      required:
        - file
      properties:
        file:
          type: string
          format: binary
          description: Document file (PDF, DOCX, or TXT)
        document_type:
          type: string
          enum: [pharmacopoeia, protocol, reference]
          default: reference
        metadata:
          type: object
          description: Additional document metadata
    DocumentUploadResponse:
      type: object
      required:
        - document_id
        - filename
        - status
      properties:
        document_id:
          type: string
          format: uuid
        filename:
          type: string
        file_size:
          type: integer
          description: File size in bytes
        status:
          type: string
          enum: [uploaded, processing, completed, failed]
        uploaded_at:
          type: string
          format: date-time
    DocumentResponse:
      type: object
      required:
        - document_id
        - filename
        - status
      properties:
        document_id:
          type: string
          format: uuid
        filename:
          type: string
        file_size:
          type: integer
        document_type:
          type: string
        status:
          type: string
          enum: [uploaded, processing, completed, failed]
        processing_stats:
          $ref: '#/components/schemas/ProcessingStats'
        uploaded_at:
          type: string
          format: date-time
        processed_at:
          type: string
          format: date-time
    ProcessingStats:
      type: object
      properties:
        total_chunks:
          type: integer
        total_pages:
          type: integer
        processing_time:
          type: number
          description: Processing time in seconds
        embedding_count:
          type: integer
    DocumentListResponse:
      type: object
      required:
        - documents
        - pagination
      properties:
        documents:
          type: array
          items:
            $ref: '#/components/schemas/DocumentResponse'
        pagination:
          $ref: '#/components/schemas/Pagination'
    ProcessingResponse:
      type: object
      required:
        - message
        - status
      properties:
        message:
          type: string
        status:
          type: string
          enum: [initiated, processing]
        estimated_completion:
          type: string
          format: date-time

    # Search Schemas
...
    Pagination:
      type: object
      required:
        - page
        - per_page
        - total
        - pages
      properties:
        page:
          type: integer
          minimum: 1
        per_page:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
        pages:
          type: integer
          minimum: 0
        has_next:
          type: boolean
        has_prev:
          type: boolean
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - ApiKeyAuth: []
  - BearerAuth: []